fix some rendering issue

+++ src/ClassicUO.Renderer/Animations/AnimationDirection.cs
@@ -2,7 +2,7 @@ namespace ClassicUO.Renderer.Animations
 {
     public struct AnimationDirection
     {
-        public long Address;
+        public uint Address;
         public byte FrameCount;
         public SpriteInfo[] SpriteInfos;
         public bool IsVerdata;
